BJS2-78398: интеграционный тест на редис

+++ src/main/java/faang/school/postservice/dto/comment/FeedCommentDto.java
@@ -14,7 +14,7 @@ public class FeedCommentDto {
     private Long postId;
     private String content;
     private Long authorId;
-    private long likes;
+    private int likes;
     private String createdAt;
     private String updatedAt;
 }
